Squashed commit of the following:
commit e9730c50766bf8506e4cc272cffa82fdb255a39c
Author: Mariocci <mp53543@fer.hr>
Date:   Sat Jan 6 19:35:07 2024 +0100

    APPProperties za master

commit 4f5e839ec56fc9cf66fa512ef49e3624d99be5e9
Author: Mariocci <mp53543@fer.hr>
Date:   Sat Jan 6 18:52:08 2024 +0100

    Automatsko slanje mailova

    Realizirano automatsko slanje mailova korisnicima.

commit 939a061622fc1a7c81fbcefe62da292d35824adf
Author: Mariocci <mp53543@fer.hr>
Date:   Sat Jan 6 17:18:26 2024 +0100

    Update ConferenceService.java

commit 39dfc4f02b84869cfd31e52d80563938ce65eafd
Author: Mariocci <mp53543@fer.hr>
Date:   Sat Jan 6 17:18:22 2024 +0100

    Create EmailService.java

commit 4fec9fd40fcbc2928132763d5e6e0f588d5fa3da
Author: Mariocci <mp53543@fer.hr>
Date:   Sat Jan 6 17:17:54 2024 +0100

    Update ConferenceService.java

commit 211e5c99b23ee58970168d73a0f6d79c0f321ae6
Merge: b004c45 6d29fe3
Author: Dario Tomšić <dt54141@fer.hr>
Date:   Sat Jan 6 17:08:57 2024 +0100

    Merge branch 'backendpromjene' of https://github.com/dominikpapes/Posterheimer into backendpromjene

commit b004c4564e5ae8094173150998c1134400940d9a
Author: Dario Tomšić <dt54141@fer.hr>
Date:   Sat Jan 6 17:05:32 2024 +0100

    osnovno slanje maila

commit 6d29fe3b365c4d8a342f9b089f3d123cf5dc3fc4
Author: Mariocci <mp53543@fer.hr>
Date:   Sat Jan 6 16:56:59 2024 +0100

    Conference service v1

commit cc0e4b79c56caad0fdb7a2cec1c491a0b0d574d8
Author: Mariocci <mp53543@fer.hr>
Date:   Sat Jan 6 11:33:25 2024 +0100

    Login update

    Dodana funkcionalnost slanja imena i prezimena na front end za login.

commit 20d3ab3a9da2c440b1bc78e6734d93956cc3fbc4
Merge: 5d11248 0d4f914
Author: Mariocci <mp53543@fer.hr>
Date:   Fri Jan 5 18:35:17 2024 +0100

    Merge branch 'backendpromjene' of https://github.com/dominikpapes/Posterheimer into backendpromjene

commit 5d11248682976fc1ebf4f10c917558d43fd2eb33
Author: Mariocci <mp53543@fer.hr>
Date:   Fri Jan 5 18:35:14 2024 +0100

    Uklonjen pristup svima za verifikaciju

commit 33a92cfc559e29aaf54451de156554662f76218d
Author: Mariocci <mp53543@fer.hr>
Date:   Fri Jan 5 18:34:49 2024 +0100

    Captcha funkcionalnost

    Dodana podrska za captcha verifikaciju "/verify" endpoint

commit 0d4f914e763b8195942879ec2715195c2015fb7c
Author: Dario Tomšić <dt54141@fer.hr>
Date:   Fri Jan 5 16:44:41 2024 +0100

    sitna promjena

commit 7d0d669547239c96f906a090dc58056cdf885b46
Author: Dario Tomšić <dt54141@fer.hr>
Date:   Fri Jan 5 16:27:03 2024 +0100

    automatsko dodijeljivanje id-a slikama

commit c344357ed5a72af8f8cfd5436e666bfb7bd65502
Author: Dario Tomšić <dt54141@fer.hr>
Date:   Fri Jan 5 15:53:11 2024 +0100

    Implementirana razmjena slika/pdfova, dodana varijabla urlPromo pokrovitelju, promjene u KorisnikGetDTO (šalje ime i prezime), manje promjene/ispravci

commit 7085fde8c269cfc811721c24fc920add44b612de
Author: Mariocci <mp53543@fer.hr>
Date:   Fri Jan 5 14:09:54 2024 +0100

    Form security autentikacija napravljena

commit 972c3c6a6ef42065225ac57b12c685c2ce12a819
Merge: 5883ff3 f9c19f1
Author: Mariocci <mp53543@fer.hr>
Date:   Thu Jan 4 22:26:47 2024 +0100

    Merge branch 'backendpromjene' of https://github.com/dominikpapes/Posterheimer into backendpromjene

commit 5883ff30e0bd8d614b0ea76bafc3ca46d365a328
Author: Mariocci <mp53543@fer.hr>
Date:   Thu Jan 4 22:26:44 2024 +0100

    Autentikacija upgrade

    Omogućen form security, koristi se JWT token sada, treba testirati i debbugirati, treba updateati sve secured anotacije.

commit f9c19f116829e4a793bf50a8ac80baafb74dbed1
Author: Dario Tomšić <dt54141@fer.hr>
Date:   Thu Jan 4 19:19:38 2024 +0100

    Promijenjen način dodijeljivanja atributa IdKonferencija

commit 7971e0dd16a28326b1fba6a69807ef2e7314155a
Author: Dario Tomšić <dt54141@fer.hr>
Date:   Thu Jan 4 18:12:10 2024 +0100

    Promijenjeni DTO i Mapperi za Konferencije (id na POST-u jos nije popravljen)

commit 1993c31d9eb67c42b6363ae4e1d235f7a7a5b27f
Author: Mariocci <mp53543@fer.hr>
Date:   Thu Jan 4 12:03:56 2024 +0100

    Manje promjene

commit 5e401e10ce9b2e3b7d8b3fc22761e6bf921d13d6
Author: Dario Tomšić <dt54141@fer.hr>
Date:   Wed Jan 3 19:20:32 2024 +0100

    male promjene

commit 71f4806a84e4fb130ab72ee338879198e7b75022
Author: Mariocci <mp53543@fer.hr>
Date:   Wed Jan 3 19:05:59 2024 +0100

    promjene

commit d89d0a80fac105998790513daf31bb2533c9c9fc
Author: Mariocci <mp53543@fer.hr>
Date:   Wed Jan 3 18:42:04 2024 +0100

    manje promjene

commit a44d19259525f10846b4486182f1673031cbaaaa
Author: Mariocci <mp53543@fer.hr>
Date:   Wed Jan 3 14:47:28 2024 +0100

    Glasovanje v2

    Popravljeno glasovanje

commit 2012a5217b3542d9cd0a399606fbcaee557b3c2e
Author: Mariocci <mp53543@fer.hr>
Date:   Wed Jan 3 11:53:32 2024 +0100

    Glasovanje impl

    Implementirano glasovanje za postere, treba sve jos detaljno istestirat

commit 48845d4920a6965902bf397343de0a4c22e49429
Author: Dario Tomšić <dt54141@fer.hr>
Date:   Tue Jan 2 20:55:26 2024 +0100

    implementirana klasa Fotografija

commit e726eda2511714f07ad65285e50c0292493ca69a
Author: Dario Tomšić <dt54141@fer.hr>
Date:   Tue Jan 2 18:27:58 2024 +0100

    Pokrovitelj Controller, manje izmjene

commit 750080d79aa5c83fb8fce7823552401668f2cabc
Author: Mariocci <mp53543@fer.hr>
Date:   Tue Jan 2 13:20:53 2024 +0100

    DTO update

    Definiran dto za postere i sve potrebno, dodan atribut glasova za poster, treba definirati glasanje i dodati metodu za glasanje kako i provjeru.

commit 9e69ef2647b8951be18e216c9fb033cae666ba53
Merge: c120bff c358860
Author: Mariocci <mp53543@fer.hr>
Date:   Tue Jan 2 12:17:31 2024 +0100

    Merge branch 'backendpromjene' of https://github.com/dominikpapes/Posterheimer into backendpromjene

commit c120bfff915f8c40bf33672ec04cac8eee46cadc
Author: Mariocci <mp53543@fer.hr>
Date:   Tue Jan 2 12:17:24 2024 +0100

    DTO update

    Definirani DTO i mapper za korisnika, updateane funkcije koje kroiste isti.

commit c35886053457b8abedff3a525744387c917c13ea
Author: Dario Tomšić <dt54141@fer.hr>
Date:   Mon Jan 1 22:29:03 2024 +0100

    Pokrovitelj DTOs

commit c921a44e71ec26a1e951c8c3e3e26a05ba545a36
Author: Dario Tomšić <dt54141@fer.hr>
Date:   Mon Jan 1 21:58:47 2024 +0100

    Pokrovitelj bez controllera

commit 4481f38b6b87562546cd3ad0f2595f6aad64d5f1
Author: Mariocci <mp53543@fer.hr>
Date:   Mon Jan 1 16:27:04 2024 +0100

    Definirani DTO za konferencije

    Napravljeni dto i mapperi za konferencije treba jos sve ostalo po tome

commit a19e9c71114943606a39a45920370e0c3f7da075
Author: Mariocci <mp53543@fer.hr>
Date:   Mon Dec 11 17:46:41 2023 +0100

    dependency za dto

commit 6661044cb33b8b6ff612bf6d7f86b344b1d342b6
Merge: 19df2ad 6c927b8
Author: Mariocci <mp53543@fer.hr>
Date:   Sat Dec 9 14:12:01 2023 +0100

    Merge branch 'master' into backendpromjene

commit 19df2ad17e8a440fe5254bfcd1840f5e5059d848
Author: Mariocci <mp53543@fer.hr>
Date:   Sat Dec 9 13:27:39 2023 +0100

    komentiranje

commit 4363c72cf77e991964fa8b1af7e375938a4bd4bf
Author: Mariocci <mp53543@fer.hr>
Date:   Sat Dec 9 13:17:09 2023 +0100

    Komentiranje i brisanje

+++ IzvorniKod/posterheimer-backend/src/main/java/com/rest/KonferencijaController.java
@@ -1,27 +1,33 @@
 package com.rest;
 import com.domain.Konferencija;
 import com.domain.Korisnik;
-import com.service.EntityMissingException;
+import com.dto.KonferencijaDTOs.KonferencijaIdDTO;
+import com.dto.KonferencijaDTOs.KonferencijaGetDTO;
+import com.dto.KonferencijaDTOs.KonferencijaPostDTO;
+import com.dto.KorisnikDTOs.KorisnikGetDTO;
+import com.mapper.KonferencijaMappers.KonferencijaGetMapper;
+import com.mapper.KonferencijaMappers.KonferencijaPostMapper;
+import com.mapper.KorisnikMappers.KorisnikGetMapper;
 import com.service.KonferencijeService;
 import com.service.KorisnikService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 //import org.springframework.security.access.annotation.Secured;
 import org.springframework.security.access.annotation.Secured;
-import org.springframework.security.core.parameters.P;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 
 import java.net.URI;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
 
+
+//Controller je prvi sloj, s njega se salje u service sloj
 @RestController
 @RequestMapping("/konferencije")
 public class KonferencijaController {
+    //definiramo service od drugih objekata ukoliko ih trebamo pozivati
     @Autowired
     private KonferencijeService konferencijeService;
 
@@ -31,63 +37,65 @@ public class KonferencijaController {
     @Autowired
     private  PasswordEncoder pwdEncoder;
 
+    //dohvacanje svih konferencija
     @GetMapping("")
-    public List<Map<String, Object>> konferencije() {
-        return konferencijeService.listAll().stream()
-                .map(konferencija -> {
-                    Map<String, Object> konferencijaMap = new HashMap<>();
-                    konferencijaMap.put("idKonferencija", konferencija.getIdKonferencija());
-                    konferencijaMap.put("imeKonferencija", konferencija.getImeKonferencija());
-                    //konferencijaMap.put("mjesto", konferencija.getMjesto());
-                    //konferencijaMap.put("datumVrijemePocetka", konferencija.getDatumVrijemePocetka());
-                    //konferencijaMap.put("datumVrijemeZavrsetka", konferencija.getDatumVrijemeZavrsetka());
-                    return konferencijaMap;
-                }).collect(Collectors.toList());
+    public List<KonferencijaGetDTO> konferencije() {
+        List<Konferencija> konferencije = konferencijeService.listAll();
+        return konferencije.stream()
+                .map(KonferencijaGetMapper::toDTO)
+                .collect(Collectors.toList());
     }
 
+    //dohvati konferenciju prema idu
     @GetMapping("/{idKonferencija}")
     @Secured({"ROLE_SUPERUSER","ROLE_ADMIN", "ROLE_USER", "ROLE_VISITOR"})
-    public Konferencija getKonferencijaById(@PathVariable("idKonferencija") Integer idKonferencija) {
-        return konferencijeService.fetch(idKonferencija);
+    public KonferencijaGetDTO getKonferencijaById(@PathVariable("idKonferencija") Integer idKonferencija) {
+        Konferencija konferencija = konferencijeService.fetch(idKonferencija);
+        return KonferencijaGetMapper.toDTO(konferencija);
     }
+    //dobavljamo listu svih usera od konferencije s poslanim IDom npr .../3/users poslat ce se sve od one s IDom 3,
+    // secured govori tko moze pristupit tome
     @GetMapping("/{idKonferencija}/users")
     @Secured({"ROLE_SUPERUSER","ROLE_ADMIN"})
-    public Set<Korisnik> getUsers(@PathVariable("idKonferencija") Integer idKonferencija) {
-        return konferencijeService.fetch(idKonferencija).getUsers();
+    public List<KorisnikGetDTO> getUsers(@PathVariable("idKonferencija") Integer idKonferencija) {
+        return konferencijeService.fetch(idKonferencija).getUsers().stream().map(KorisnikGetMapper::toDTO)
+                .collect(Collectors.toList());
     }
-    /*@GetMapping("/{idKonferencija}/{genericPassword}")
-    public boolean loginGeneric(@PathVariable("idKonferencija") Integer idKonferencija, @PathVariable("genericPassword") String genericPassword){
-        try{
-        Konferencija saved=konferencijeService.fetch(idKonferencija);
-        return saved.getGenericPassword().equals(genericPassword);}
-        catch (EntityMissingException exception){
-            throw new EntityMissingException(Konferencija.class,idKonferencija);
-        }
-    }*/
 
+    //dto
     @PostMapping("")
   //@Secured("ROLE_SUPERUSER")
-    public ResponseEntity<Konferencija> createKonferencija(@RequestBody Konferencija konferencija) {
+    public ResponseEntity<Konferencija> createKonferencija(@RequestBody KonferencijaPostDTO konferencijaPostDTO) {
+        Konferencija konferencija = KonferencijaPostMapper.toEntity(konferencijaPostDTO);
+        konferencija.setIdKonferencija(-1);
         Korisnik tempKorisnik = new Korisnik(konferencija.getGenericUsername(), pwdEncoder.encode(konferencija.getGenericPassword()), konferencija.getIdKonferencija().toString(), konferencija.getImeKonferencija(), false, true);
         konferencija.setGenericPassword(tempKorisnik.getLozinka());
         tempKorisnik.setKonferencija(konferencija);
         Konferencija saved = konferencijeService.createKonferencija(konferencija);
         korisnikService.createKorisnik(tempKorisnik);
+        tempKorisnik=new Korisnik(konferencijaPostDTO.getAdminUsername(), konferencijaPostDTO.getAdminPassword(), null,null,true,false);
+        korisnikService.createKorisnik(tempKorisnik);
         return ResponseEntity.created(URI.create("/konferencije/" + saved.getIdKonferencija())).body(saved);
     }
 
-    @PutMapping("/{id}")
-  //  @Secured("ROLE_SUPERUSER")
+
+    /*@PutMapping("/{id}")
+  //@Secured("ROLE_SUPERUSER")
     public Konferencija konferencija(@PathVariable("id") Integer id, @RequestBody Konferencija konferencija) {
         if (!(konferencija.getIdKonferencija().equals(id)))
             throw new IllegalArgumentException("");
         return konferencijeService.updateKonferencija(konferencija,id);
-    }
+    }*/
 
-
-    @DeleteMapping("/{id}")
- //   @Secured("ROLE_SUPERUSER")
-    public Konferencija deleteKonferencija(@PathVariable("id") Integer idKonferencija) {
-        return konferencijeService.deleteKonferencija(idKonferencija);
+    //dto
+    @DeleteMapping("")
+  //@Secured("ROLE_SUPERUSER")
+    public ResponseEntity<?> deleteKonferencija(@RequestBody KonferencijaIdDTO konferencijaIdDTO) {
+        Integer idKonferencija = konferencijaIdDTO.getIdKonferencija();
+        if (idKonferencija == null) {
+            return ResponseEntity.badRequest().body("ID must not be null");
+        }
+        konferencijeService.deleteKonferencija(idKonferencija);
+        return ResponseEntity.noContent().build();
     }
 }
